# Sources/NetCoreTestApp/NetCoreTestApp.Api/swagger.json
{"swagger":"2.0","info":{"version":"v1","title":"ToDo API","description":"A simple example ASP.NET Core Web API","termsOfService":"None","contact":{"name":"Yauheniya Smolnikava","email":""},"license":{"name":"Use under LICX","url":"https://example.com/license"}},"paths":{"/api/Categories":{"get":{"tags":["Categories"],"summary":"This methid allows to get all available catgories","operationId":"Get","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[],"responses":{"200":{"description":"Success","schema":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/Categories"}}}}}},"/api/Categories/images/{id}":{"get":{"tags":["Categories"],"summary":"This method allows to get image for the specified category","operationId":"GetImage","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","description":"identifier of category","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success"}}},"post":{"tags":["Categories"],"summary":"This method allows to update image for the specified category","operationId":"Upload","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"image","in":"body","description":"image object","required":false,"schema":{"$ref":"#/definitions/ImageUpload"}},{"name":"id","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/api/Products":{"get":{"tags":["Products"],"summary":"This method returns all available products","operationId":"Get","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[],"responses":{"200":{"description":"Success","schema":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/Products"}}}}},"post":{"tags":["Products"],"summary":"This method allows to create new product based on passed param","operationId":"Post","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"product","in":"body","description":"Filled product object","required":false,"schema":{"$ref":"#/definitions/Products"}}],"responses":{"200":{"description":"Success"}}}},"/api/Products/{id}":{"put":{"tags":["Products"],"summary":"This method allows to update existing product","operationId":"Put","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"id","in":"path","description":"id of existing product","required":true,"type":"integer","format":"int32"},{"name":"product","in":"body","description":"product with updated data","required":false,"schema":{"$ref":"#/definitions/Products"}}],"responses":{"200":{"description":"Success"}}},"delete":{"tags":["Products"],"summary":"This method allows to delete exisiting product","operationId":"Delete","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","description":"identifier of product that need to be deleted","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success"}}}}},"definitions":{"Categories":{"type":"object","properties":{"categoryId":{"format":"int32","type":"integer"},"categoryName":{"type":"string"},"description":{"type":"string"},"picture":{"format":"byte","type":"string"},"products":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/Products"}}}},"Products":{"required":["productName"],"type":"object","properties":{"productId":{"format":"int32","type":"integer"},"productName":{"maxLength":100,"minLength":0,"type":"string"},"supplierId":{"format":"int32","type":"integer"},"categoryId":{"format":"int32","type":"integer"},"quantityPerUnit":{"type":"string"},"unitPrice":{"format":"double","type":"number"},"unitsInStock":{"format":"int32","type":"integer"},"unitsOnOrder":{"format":"int32","type":"integer"},"reorderLevel":{"format":"int32","maximum":35.0,"minimum":0.0,"type":"integer"},"discontinued":{"type":"boolean"},"category":{"$ref":"#/definitions/Categories"},"supplier":{"$ref":"#/definitions/Suppliers"},"orderDetails":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/OrderDetails"}}}},"Suppliers":{"type":"object","properties":{"supplierId":{"format":"int32","type":"integer"},"companyName":{"type":"string"},"contactName":{"type":"string"},"contactTitle":{"type":"string"},"address":{"type":"string"},"city":{"type":"string"},"region":{"type":"string"},"postalCode":{"type":"string"},"country":{"type":"string"},"phone":{"type":"string"},"fax":{"type":"string"},"homePage":{"type":"string"},"products":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/Products"}}}},"OrderDetails":{"type":"object","properties":{"orderId":{"format":"int32","type":"integer"},"productId":{"format":"int32","type":"integer"},"unitPrice":{"format":"double","type":"number"},"quantity":{"format":"int32","type":"integer"},"discount":{"format":"float","type":"number"},"order":{"$ref":"#/definitions/Orders"},"product":{"$ref":"#/definitions/Products"}}},"Orders":{"type":"object","properties":{"orderId":{"format":"int32","type":"integer"},"customerId":{"type":"string"},"employeeId":{"format":"int32","type":"integer"},"orderDate":{"format":"date-time","type":"string"},"requiredDate":{"format":"date-time","type":"string"},"shippedDate":{"format":"date-time","type":"string"},"shipVia":{"format":"int32","type":"integer"},"freight":{"format":"double","type":"number"},"shipName":{"type":"string"},"shipAddress":{"type":"string"},"shipCity":{"type":"string"},"shipRegion":{"type":"string"},"shipPostalCode":{"type":"string"},"shipCountry":{"type":"string"},"customer":{"$ref":"#/definitions/Customers"},"employee":{"$ref":"#/definitions/Employees"},"shipViaNavigation":{"$ref":"#/definitions/Shippers"},"orderDetails":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/OrderDetails"}}}},"Customers":{"type":"object","properties":{"customerId":{"type":"string"},"companyName":{"type":"string"},"contactName":{"type":"string"},"contactTitle":{"type":"string"},"address":{"type":"string"},"city":{"type":"string"},"region":{"type":"string"},"postalCode":{"type":"string"},"country":{"type":"string"},"phone":{"type":"string"},"fax":{"type":"string"},"customerCustomerDemo":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/CustomerCustomerDemo"}},"orders":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/Orders"}}}},"Employees":{"type":"object","properties":{"employeeId":{"format":"int32","type":"integer"},"lastName":{"type":"string"},"firstName":{"type":"string"},"title":{"type":"string"},"titleOfCourtesy":{"type":"string"},"birthDate":{"format":"date-time","type":"string"},"hireDate":{"format":"date-time","type":"string"},"address":{"type":"string"},"city":{"type":"string"},"region":{"type":"string"},"postalCode":{"type":"string"},"country":{"type":"string"},"homePhone":{"type":"string"},"extension":{"type":"string"},"photo":{"format":"byte","type":"string"},"notes":{"type":"string"},"reportsTo":{"format":"int32","type":"integer"},"photoPath":{"type":"string"},"reportsToNavigation":{"$ref":"#/definitions/Employees"},"employeeTerritories":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/EmployeeTerritories"}},"inverseReportsToNavigation":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/Employees"}},"orders":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/Orders"}}}},"Shippers":{"type":"object","properties":{"shipperId":{"format":"int32","type":"integer"},"companyName":{"type":"string"},"phone":{"type":"string"},"orders":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/Orders"}}}},"CustomerCustomerDemo":{"type":"object","properties":{"customerId":{"type":"string"},"customerTypeId":{"type":"string"},"customer":{"$ref":"#/definitions/Customers"},"customerType":{"$ref":"#/definitions/CustomerDemographics"}}},"EmployeeTerritories":{"type":"object","properties":{"employeeId":{"format":"int32","type":"integer"},"territoryId":{"type":"string"},"employee":{"$ref":"#/definitions/Employees"},"territory":{"$ref":"#/definitions/Territories"}}},"CustomerDemographics":{"type":"object","properties":{"customerTypeId":{"type":"string"},"customerDesc":{"type":"string"},"customerCustomerDemo":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/CustomerCustomerDemo"}}}},"Territories":{"type":"object","properties":{"territoryId":{"type":"string"},"territoryDescription":{"type":"string"},"regionId":{"format":"int32","type":"integer"},"region":{"$ref":"#/definitions/Region"},"employeeTerritories":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/EmployeeTerritories"}}}},"Region":{"type":"object","properties":{"regionId":{"format":"int32","type":"integer"},"regionDescription":{"type":"string"},"territories":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/Territories"}}}},"ImageUpload":{"type":"object","properties":{"file":{"$ref":"#/definitions/IFormFile"},"categoryId":{"format":"int32","type":"integer"}}},"IFormFile":{"type":"object","properties":{"contentType":{"type":"string","readOnly":true},"contentDisposition":{"type":"string","readOnly":true},"headers":{"type":"object","additionalProperties":{"uniqueItems":false,"type":"array","items":{"type":"string"}},"readOnly":true},"length":{"format":"int64","type":"integer","readOnly":true},"name":{"type":"string","readOnly":true},"fileName":{"type":"string","readOnly":true}}}}}